Introducing Task Lock in Task Menu
In situations where users frequently run the app in the background, and open multiple apps, there may be instances where a specific app should not be cleared when performing a 'Clear All' or 'Dismiss' action. This new feature ensures that a selected running task or app remains in the Recent View and will not be removed.

Change-Id: Idc8cb1e984f2b0dc169c38d86513107eee3ec26e